requirements.txt
blinker==1.9.0
click==8.1.8
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.3
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.5.0
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.10.0
numba==0.60.0
numpy==2.0.2
nvidia-nccl-cu12==2.24.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
shap==0.46.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
Werkzeug==3.1.3
xgboost==2.1.3
